requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
build==1.1.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
Chroma==0.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
dashscope==1.16.0
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
deepdiff==6.7.1
Deprecated==1.2.14
distro==1.9.0
docx==0.2.4
effdet==0.4.1
emoji==2.10.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.1
filetype==1.2.0
flatbuffers==24.3.7
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.28.2
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.21.4
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.3.2
iopath==0.1.10
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.32
langchain-google-genai==0.0.11
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langdetect==1.0.9
langsmith==0.1.31
layoutparser==0.3.4
lxml==5.1.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.3
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
ollama==0.1.7
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==1.14.2
opencv-python==4.9.0.80
openpyxl==3.1.2
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
ordered-set==4.1.0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
pikepdf==8.13.0
pillow==10.2.0
pillow_heif==0.15.0
portalocker==2.8.2
posthog==3.5.0
proto-plus==1.23.0
protobuf==4.25.3
pulsar-client==3.4.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pyparsing==3.1.2
pypdf==4.1.0
pypdfium2==4.28.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-iso639==2024.2.7
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
pywin32==306
PyYAML==6.0.1
rapidfuzz==3.6.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
safetensors==0.4.2
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
starlette==0.36.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.6.0
timm==0.9.16
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
transformers==4.38.2
typer==0.9.0
types-requests==2.31.0.20240311
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
unstructured==0.12.5
unstructured-client==0.22.0
unstructured-inference==0.7.24
unstructured.pytesseract==0.3.12
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
xlrd==2.0.1
yarl==1.9.4
zipp==3.18.1
